organized the functions and class
