+++ main.py
@@ -5,6 +5,8 @@ import httpx
 from fastapi import FastAPI, Request, Response
 from dotenv import load_dotenv
 import os
+from schemas.helper import *
+from schemas.schema import WebhookRequestData, UserInfo
 
 app = FastAPI()
 load_dotenv()
@@ -13,166 +15,7 @@ API_URL = os.getenv('META_API_URL')
 ACCESS_TOKEN = os.getenv('PAGE_ACCESS_TOKEN')
 VERIFY_TOKEN = os.getenv('VERIFY_TOKEN')
 
-
-class WebhookRequestData(BaseModel):
-    object: str = ""
-    entry: List = []
-
 # Helpers
-
-
-async def send_message(
-    url: str,
-    page_access_token: str,
-    recipient_id: str,
-    message_text: str,
-    message_type: str = "UPDATE",
-):
-    try:
-        response = httpx.post(
-            url=url,
-            params={"access_token": page_access_token},
-            headers={"Content-Type": "application/json"},
-            json={
-                "recipient": {"id": recipient_id},
-                "message": {"text": message_text},
-                "messaging_type": message_type,
-            },
-        )
-    except:
-        # This is an http error
-        response.raise_for_status()
-
-# Helpers
-
-
-async def send_button_message(
-    url: str,
-    page_access_token: str,
-    recipient_id: str,
-    message_type: str = "RESPONSE",
-):
-    try:
-        response = httpx.post(
-            url=url,
-            params={"access_token": page_access_token},
-            headers={"Content-Type": "application/json"},
-            json={
-                "recipient": {"id": recipient_id},
-                "message": {
-                    "text": "What do you want to know?",
-                    "quick_replies": [
-                        {
-                            "content_type": "text",
-                            "title": "Weather",
-                            "payload": "I want to know weather"
-                        }, {
-                            "content_type": "text",
-                            "title": "temperature",
-                            "payload": "I want to know temperature"
-                        }
-                    ]
-                },
-                "messaging_type": message_type,
-            },
-        )
-    except:
-        # This is an http error
-        response.raise_for_status()
-
-
-async def send_next_message(
-    url: str,
-    page_access_token: str,
-    recipient_id: str,
-    message_type: str = "RESPONSE",
-):
-    try:
-        response = httpx.post(
-            url=url,
-            params={"access_token": page_access_token},
-            headers={"Content-Type": "application/json"},
-            json={
-                "recipient": {"id": recipient_id},
-                "message": {
-                    "attachment": {
-                        "type": "template",
-                        "payload": {
-                            "template_type": "button",
-                            "text": "What do you want to do next?",
-                            "buttons": [
-                                {
-                                    "type": "web_url",
-                                    "url": "https://www.messenger.com",
-                                    "title": "Visit Messenger"
-                                },
-                                {
-                                    "type": "web_url",
-                                    "url": "https://www.youtube.com",
-                                    "title": "Visit Youtube"
-                                },
-                            ]
-                        }
-                    }
-                }
-            }
-        )
-    except:
-        # This is an http error
-        response.raise_for_status()
-
-
-async def send_home_message(
-    url: str,
-    page_access_token: str,
-    recipient_id: str,
-    message_type: str = "RESPONSE",
-):
-    try:
-        home_response = httpx.post(
-            url=url,
-            params={"access_token": page_access_token},
-            headers={"Content-Type": "application/json"},
-            json={
-                "recipient": {
-                    "id": recipient_id
-                },
-                "message": {
-                    "attachment": {
-                        "type": "template",
-                        "payload": {
-                                "template_type": "generic",
-                                "elements": [
-                                    {
-                                        "title": "Welcome!",
-                                        "image_url": "https://media.cntraveler.com/photos/60480c67ff9cba52f2a91899/16:9/w_2560%2Cc_limit/01-velo-header-seattle-needle.jpg",
-                                        "subtitle": "Your BEST Seattle local guide!",
-                                        "default_action": {
-                                            "type": "web_url",
-                                            "url": "https://github.com/GaryHo34/SeattleBot",
-                                            "webview_height_ratio": "tall",
-                                        },
-                                        "buttons": [
-                                            {
-                                                "type": "web_url",
-                                                "url": "https://github.com/GaryHo34/SeattleBot",
-                                                "title": "Our GitHub Page"
-                                            }, {
-                                                "type": "postback",
-                                                "title": "Start Chatting",
-                                                "payload": "DEVELOPER_DEFINED_PAYLOAD"
-                                            }
-                                        ]
-                                    }
-                                ]
-                        }
-                    }
-                }
-            }
-        )
-    except:
-        # This is an http error
-        home_response.raise_for_status()
 
 
 @ app.get("/")
@@ -200,18 +43,16 @@ async def webhook(data: WebhookRequestData):
                 message = event.get("message")
                 sender_id = event["sender"]["id"]
 
+                user = UserInfo(url=API_URL,
+                                page_access_token=ACCESS_TOKEN,
+                                recipient_id=sender_id)
+
                 if message['text'] == "weather":
-                    await send_button_message(url=API_URL,
-                                              page_access_token=ACCESS_TOKEN,
-                                              recipient_id=sender_id)
+                    await send_button_message(user)
                 elif message['text'] == "next":
-                    await send_next_message(url=API_URL,
-                                            page_access_token=ACCESS_TOKEN,
-                                            recipient_id=sender_id)
+                    await send_next_message(user)
                 else:
-                    await send_home_message(url=API_URL,
-                                            page_access_token=ACCESS_TOKEN,
-                                            recipient_id=sender_id)
+                    await send_home_message(user)
 
     print("PK")
     return Response(content="ok")
